Refactor FormatConverter: move SedtrailsData class to its own module
